+++ src/modules/react-redux-hash-router4/view/components/index.js
@@ -1,0 +1,7 @@
+import _Link from './Link'
+import _IndexLink from './IndexLink'
+import _Router from './Router'
+
+export const Link = _Link
+export const Router = _Router
+export const IndexLink = _IndexLink